21/03/2022 23:03

+++ src/Voertuig.h
@@ -13,7 +13,7 @@ using namespace std;
 class Voertuig{
 private:
     string baan;
-    int positie;
+    double positie;
     int lengte;
     double snelheid;
     double versnelling;
@@ -28,9 +28,9 @@ public:
     double getSnelheid() const;
     void setSnelheid(double snelheid);
     const string &getBaan();
-    int getPositie();
+    double getPositie();
     void setBaan(string const &name);
-    void setPositie(int number);
+    void setPositie(double number);
 
 };
